refactor: cohesive state of InflationRate chart
